requirements.txt
# ─── Core Django stack ─────────────────────────────────────────────
Django==5.1.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-decouple==3.8
whitenoise==6.6.0         # serve static files in DEBUG=False
gunicorn==22.0.0          # prod WSGI server
python-dotenv==1.0.1
# ─── ML / Prediction ───────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.2
yfinance==0.2.38
tensorflow==2.16.1        # or tf-cpu to slim down
#keras==2.16.1

# ─── Optional / Upcoming Milestones ────────────────────────────────
python-telegram-bot[rate-limiter]==22.1 # Telegram bot
redis==5.0.4              # Celery broker / rate‑limit cache
celery==5.3.6
stripe==9.4.0             # paid memberships

# ─── Utility ───────────────────────────────────────────────────────
requests==2.32.3
pillow==10.4.0
